fix add search
