+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDaoStorage.java
@@ -1,6 +1,7 @@
 
 package ru.yandex.practicum.filmorate.storage.film;
 
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.List;
@@ -41,5 +42,7 @@ public interface FilmDaoStorage {
     List<Film> getSearchFilmsForDirector(String query);
 
     List<Film> getSearchFilmsForTitleAndDirector(String query);
+
+    Director updateDirector(Director director);
 }
