+++ src/main/resources/schema.sql
@@ -1,101 +1,144 @@
-CREATE TABLE IF NOT EXISTS RATINGS
+drop table if exists DIRECTORS cascade;
+drop table if exists film_director cascade;
+drop table if exists FILMS cascade;
+drop table if exists FILMS_GENRES cascade;
+drop table if exists FILMS_LIKES cascade;
+drop table if exists FRIENDS cascade;
+drop table if exists GENRES cascade;
+drop table if exists RATINGS cascade;
+drop table if exists REVIEW cascade;
+drop table if exists REVIEW_LIKES cascade;
+drop table if exists USERS cascade;
+
+
+create table DIRECTORS
 (
-    RATING_ID   INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    RATING_NAME varchar(10), -- NOT NULL ,
-    CONSTRAINT uc_rating_name UNIQUE (RATING_NAME)
+    DIRECTOR_ID   INTEGER auto_increment
+        primary key,
+    DIRECTOR_NAME CHARACTER VARYING(60)
+        constraint UC_DIRECTOR_NAME
+            unique
 );
 
-CREATE TABLE IF NOT EXISTS GENRES
+create table GENRES
 (
-    GENRE_ID LONG GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    NAME     VARCHAR(30),-- NOT NULL,
-    CONSTRAINT genre_name UNIQUE (NAME)
+    GENRE_ID BIGINT auto_increment
+        primary key,
+    NAME     CHARACTER VARYING(30)
+        constraint GENRE_NAME
+            unique
 );
 
-CREATE TABLE IF NOT EXISTS DIRECTORS
+create table RATINGS
 (
-    DIRECTOR_ID INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    DIRECTOR_NAME VARCHAR(60),
-    CONSTRAINT uc_director_name UNIQUE (DIRECTOR_NAME)
+    RATING_ID   INTEGER auto_increment
+        primary key,
+    RATING_NAME CHARACTER VARYING(10)
+        constraint UC_RATING_NAME
+            unique
 );
 
-
-
-CREATE TABLE IF NOT EXISTS FILMS
+create table FILMS
 (
-    FILM_ID      LONG GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    NAME         VARCHAR(150),
-
-    DESCRIPTION  VARCHAR(500) NOT NULL,
-    RELEASE_DATE DATE         NOT NULL,
-    DURATION     INTEGER      NOT NULL,
-    RATING_ID    INTEGER REFERENCES RATINGS (RATING_ID),
-    DIRECTOR_ID INTEGER REFERENCES DIRECTORS (DIRECTOR_ID)
-    --CONSTRAINT fk_film_rating_id FOREIGN KEY (RATING_ID) REFERENCES RATINGS (RATING_ID)
+    FILM_ID      BIGINT auto_increment
+        primary key,
+    NAME         CHARACTER VARYING(150),
+    DESCRIPTION  CHARACTER VARYING(500) not null,
+    RELEASE_DATE DATE                   not null,
+    DURATION     INTEGER                not null,
+    RATING_ID    INTEGER
+        references RATINGS,
+    DIRECTOR_ID  INTEGER
+        references DIRECTORS
 );
 
-CREATE TABLE IF NOT EXISTS film_director (
-                                             FILM_ID LONG REFERENCES FILMS(FILM_ID),
-                                             DIRECTOR_ID INTEGER REFERENCES DIRECTORS(DIRECTOR_ID),
-                                             PRIMARY KEY (FILM_ID, DIRECTOR_ID)
-);
-
-CREATE TABLE IF NOT EXISTS USERS
+create table FILMS_GENRES
 (
-    USER_ID   LONG GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    EMAIL     VARCHAR(50) NOT NULL,
-    LOGIN     VARCHAR(50) NOT NULL,
-    USER_NAME VARCHAR(50) NOT NULL,
-    BIRTHDAY  DATE        NOT NULL,
-    CONSTRAINT user_email UNIQUE (EMAIL),
-    CONSTRAINT user_login UNIQUE (LOGIN)
+    FILM_ID  BIGINT not null,
+    GENRE_ID BIGINT not null,
+    constraint GENRE_FILM
+        primary key (FILM_ID, GENRE_ID),
+    constraint GENRE_FILM_FILM_ID
+        foreign key (FILM_ID) references FILMS
+            on delete cascade,
+    constraint GENRE_FILM_GENRE_ID
+        foreign key (GENRE_ID) references GENRES
 );
 
-CREATE TABLE IF NOT EXISTS FILMS_LIKES
+create table FILM_DIRECTOR
 (
-    FILM_ID LONG NOT NULL,
-    USER_ID LONG NOT NULL,
-    CONSTRAINT film_like PRIMARY KEY (FILM_ID, USER_ID),
-    CONSTRAINT film_like_film_id FOREIGN KEY (FILM_ID) REFERENCES FILMS (FILM_ID)
-        ON DELETE CASCADE,
-    CONSTRAINT film_like_user_id FOREIGN KEY (USER_ID) REFERENCES USERS (USER_ID)
+    FILM_ID     BIGINT  not null
+        references FILMS,
+    DIRECTOR_ID INTEGER not null,
+    constraint FILM_DIRECTOR_DIRECTORS_DIRECTOR_ID_FK
+        foreign key (DIRECTOR_ID) references DIRECTORS
 );
 
-CREATE TABLE IF NOT EXISTS FRIENDS
+create table USERS
 (
-    USER_ID   LONG NOT NULL,
-    FRIEND_ID LONG NOT NULL,
-    CONSTRAINT friendship PRIMARY KEY (USER_ID, FRIEND_ID),
-    CONSTRAINT friendship_user_id1 FOREIGN KEY (USER_ID) REFERENCES users (USER_ID)
-        ON DELETE CASCADE,
-    CONSTRAINT friendship_user_id2 FOREIGN KEY (FRIEND_ID) REFERENCES users (USER_ID)
-        ON DELETE CASCADE
+    USER_ID   BIGINT auto_increment
+        primary key,
+    EMAIL     CHARACTER VARYING(50) not null
+        constraint USER_EMAIL
+            unique,
+    LOGIN     CHARACTER VARYING(50) not null
+        constraint USER_LOGIN
+            unique,
+    USER_NAME CHARACTER VARYING(50) not null,
+    BIRTHDAY  DATE                  not null
 );
 
-CREATE TABLE IF NOT EXISTS FILMS_GENRES
+create table FILMS_LIKES
 (
-    FILM_ID  LONG NOT NULL,
-    GENRE_ID LONG NOT NULL,
-    CONSTRAINT GENRE_FILM PRIMARY KEY (FILM_ID, GENRE_ID),
-    CONSTRAINT GENRE_FILM_film_id FOREIGN KEY (FILM_ID) REFERENCES FILMS (FILM_ID)
-        ON DELETE CASCADE,
-    CONSTRAINT GENRE_FILM_genre_id FOREIGN KEY (GENRE_ID) REFERENCES GENRES (GENRE_ID)
+    FILM_ID BIGINT not null,
+    USER_ID BIGINT not null,
+    constraint FILM_LIKE
+        primary key (FILM_ID, USER_ID),
+    constraint FILM_LIKE_FILM_ID
+        foreign key (FILM_ID) references FILMS
+            on delete cascade,
+    constraint FILM_LIKE_USER_ID
+        foreign key (USER_ID) references USERS
 );
 
-CREATE TABLE IF NOT EXISTS REVIEW
+create table FRIENDS
 (
-    REVIEW_ID LONG GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    REVIEW_CONTENT VARCHAR(1000) NOT NULL,
-    REVIEW_IS_POSITIVE BOOLEAN NOT NULL,
-    USER_ID LONG NOT NULL REFERENCES USERS (USER_ID) ON DELETE CASCADE,
-    FILM_ID LONG NOT NULL REFERENCES FILMS (FILM_ID) ON DELETE CASCADE,
-    REVIEW_USEFUL INTEGER
+    USER_ID   BIGINT not null,
+    FRIEND_ID BIGINT not null,
+    constraint FRIENDSHIP
+        primary key (USER_ID, FRIEND_ID),
+    constraint FRIENDSHIP_USER_ID1
+        foreign key (USER_ID) references USERS
+            on delete cascade,
+    constraint FRIENDSHIP_USER_ID2
+        foreign key (FRIEND_ID) references USERS
+            on delete cascade
 );
 
-CREATE TABLE IF NOT EXISTS REVIEW_LIKES
+create table REVIEW
 (
-    REVIEW_ID LONG NOT NULL REFERENCES REVIEW (REVIEW_ID) ON DELETE CASCADE,
-    USER_ID LONG NOT NULL REFERENCES USERS (USER_ID) ON DELETE CASCADE,
-    IS_LIKE BOOLEAN
+    REVIEW_ID          BIGINT auto_increment
+        primary key,
+    REVIEW_CONTENT     CHARACTER VARYING(1000) not null,
+    REVIEW_IS_POSITIVE BOOLEAN                 not null,
+    USER_ID            BIGINT                  not null
+        references USERS
+            on delete cascade,
+    FILM_ID            BIGINT                  not null
+        references FILMS
+            on delete cascade,
+    REVIEW_USEFUL      INTEGER
 );
+
+create table REVIEW_LIKES
+(
+    REVIEW_ID BIGINT not null
+        references REVIEW
+            on delete cascade,
+    USER_ID   BIGINT not null
+        references USERS
+            on delete cascade,
+    IS_LIKE   BOOLEAN
+);
+
